Code to debug 2
