+++ src/components/Rocket.js
@@ -1,7 +1,26 @@
-import React from "react";
+import React, { useEffect } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
+import { getRockets } from '../store/actions';
+import { rocketsActions } from '../store/rockets/rockets-slice';
 
-const Rocket =(rocket)=> (
-    <div className="rocket-container">
+const Rockets = () => {
+  const dispatch = useDispatch();
+  const rockets = useSelector((state) => state.rockets.rockets) || [];
+  console.log(rockets.reserved);
+  useEffect(() => {
+    dispatch(getRockets());
+  }, []);
+
+  const toggleReserve = (e) => {
+    e.preventDefault();
+    const { id } = e.target;
+    dispatch(rocketsActions.rocketReserved(id));
+  };
+
+  return (
+    <>
+      {rockets.map((rocket) => (
+        <div key={rocket.id} className="rocket-container">
           <div className="rocket-img">
             <img src={rocket.flickr_image} alt="rocket" />
           </div>
@@ -11,11 +30,14 @@ const Rocket =(rocket)=> (
               <span className="desactive">Reserved</span>
               {rocket.description}
             </p>
-            <button id={rocket} type="button" onClick={toggleReserve}>
+            <button id={rocket.id} type="button" onClick={toggleReserve}>
               Reserve Rocket
             </button>
           </div>
         </div>
-)
+      ))}
+    </>
+  );
+};
 
-export default Rocket
+export default Rockets;
